+++ test/rest.spec.js
@@ -2,6 +2,7 @@ import {Config} from '../src/config';
 import {Rest} from '../src/rest';
 import {Container} from 'aurelia-dependency-injection';
 import {InjectTest} from './resources/inject-test';
+import {buildQueryString} from 'aurelia-path';
 
 let container = new Container();
 let config    = container.get(Config);
@@ -9,26 +10,43 @@ let baseUrls  = {
   jsonplaceholder: 'http://jsonplaceholder.typicode.com/',
   api            : 'http://127.0.0.1:1927/'
 };
+let options = {
+  headers: {
+    'Content-Type' : 'application/x-www-form-urlencoded',
+    'Authorization': 'Bearer aToken'
+  }
+};
+let jsonOptions = {
+  headers: {
+    'Content-Type': 'application/vnd.api+json',
+    'Accept': 'application/vnd.api+json'
+  }
+};
 
 config.registerEndpoint('api', baseUrls.api);
 config.registerEndpoint('jsonplaceholder', baseUrls.jsonplaceholder);
 config.registerEndpoint('form', baseUrls.api, null);
+config.registerEndpoint('urlencoded', baseUrls.api, options);
+config.registerEndpoint('fetchConfig', fetchConfig => {
+  fetchConfig
+    .withBaseUrl(baseUrls.api)
+    .withDefaults(jsonOptions);
+  });
 
 let criteria = {user: 'john', comment: 'last'};
 let criteriaWithArray = {sort: ['first', 'last']};
 let body = {message: 'some'};
-let options = {
-  headers: {
-    'Content-Type': 'application/x-www-form-urlencoded',
-    'Authorization': 'Bearer aToken'
-  }
-};
 
 describe('Rest', function() {
   describe('.find()', function() {
     it('Should find results for multiple endpoints.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
+
+      let responseOutput = {
+        response: null
+      };
 
       expect(injectTest.apiEndpoint instanceof Rest).toBe(true);
       expect(injectTest.jsonplaceholderEndpoint instanceof Rest).toBe(true);
@@ -94,6 +112,10 @@ describe('Rest', function() {
             expect(y.query.sort[0]).toBe(criteriaWithArray.sort[0]);
             expect(y.query.sort[1]).toBe(criteriaWithArray.sort[1]);
             expect(y.originalUrl).toBe(encodeURI(`/posts?sort[]=${criteriaWithArray.sort[0]}&sort[]=${criteriaWithArray.sort[1]}`));
+          }),
+        injectTest.apiEndpoint.find('posts', undefined, null, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
           })
       ]).then(x => {
         done();
@@ -101,6 +123,7 @@ describe('Rest', function() {
     });
     it('Should find with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -118,6 +141,10 @@ describe('Rest', function() {
   describe('.findOne()', function() {
     it('Should find with id, criteria and options.', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -146,13 +173,18 @@ describe('Rest', function() {
             expect(y.query.sort[0]).toBe(criteriaWithArray.sort[0]);
             expect(y.query.sort[1]).toBe(criteriaWithArray.sort[1]);
             expect(y.originalUrl).toBe(encodeURI(`/posts/id/?sort[]=${criteriaWithArray.sort[0]}&sort[]=${criteriaWithArray.sort[1]}`));
-          })
+          }),
+        injectTest.apiEndpoint.findOne('posts/', 'id', criteriaWithArray, null, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
+        })
       ]).then(x => {
         done();
       });
     });
     it('Should findOne with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -165,6 +197,73 @@ describe('Rest', function() {
             expect(y.originalUrl).toBe(`/posts/id/?sort=${criteriaWithArray.sort[0]}&sort=${criteriaWithArray.sort[1]}`);
           })
       ]).then(x => done());
+    });
+  });
+
+  describe('.find()', function() {
+    it('Should find with id and criteria using date objects.', function(done) {
+      let injectTest = container.get(InjectTest);
+      let dateCriteria = { date: new Date() };
+
+      injectTest.apiEndpoint.findOne('posts', 'id', dateCriteria)
+        .then(y => {
+          expect(y.method).toBe('GET');
+          expect(y.path).toBe('/posts/id');
+          expect(y.query.date).toBe(dateCriteria.date.toString());
+          expect(y.originalUrl).toBe('/posts/id?date='+encodeURIComponent(dateCriteria.date).toString())
+        }).then(done);
+    });
+
+    it('Should find with criteria using id and date objects.', function(done) {
+      let injectTest = container.get(InjectTest);
+      let dateCriteria = { id: 'id', date: new Date() };
+
+      injectTest.apiEndpoint.findOne('posts', dateCriteria)
+        .then(y => {
+          expect(y.method).toBe('GET');
+          expect(y.path).toBe('/posts');
+          expect(y.query.id).toBe('id');
+          expect(y.query.date).toBe(dateCriteria.date.toString());
+          expect(y.originalUrl).toBe('/posts?date='+encodeURIComponent(dateCriteria.date).toString()+'&id=id')
+        }).then(done);
+    });
+
+    it('Should find with criteria using number objects.', function(done) {
+      let injectTest = container.get(InjectTest);
+      let numCriteria = { num: Number(-1.01) };
+
+      injectTest.apiEndpoint.findOne('posts/', 'id', numCriteria)
+        .then(y => {
+          expect(y.method).toBe('GET');
+          expect(y.path).toBe('/posts/id/');
+          expect(y.query.num).toBe(numCriteria.num.toString());
+        }).then(done);
+    });
+  });
+
+  describe('.find()', function() {
+    it('Should find with criteria using date objects.', function(done) {
+      let injectTest = container.get(InjectTest);
+      let dateCriteria = { date: new Date() };
+
+      injectTest.apiEndpoint.findOne('posts', 'id', dateCriteria)
+        .then(y => {
+          expect(y.method).toBe('GET');
+          expect(y.path).toBe('/posts/id');
+          expect(y.query.date).toBe(dateCriteria.date.toString());
+        }).then(done);
+    });
+
+    it('Should find with criteria using number objects.', function(done) {
+      let injectTest = container.get(InjectTest);
+      let numCriteria = { num: Number(-1.01) };
+
+      injectTest.apiEndpoint.findOne('posts/', 'id', numCriteria)
+        .then(y => {
+          expect(y.method).toBe('GET');
+          expect(y.path).toBe('/posts/id/');
+          expect(y.query.num).toBe(numCriteria.num.toString());
+        }).then(done);
     });
   });
 
@@ -191,6 +290,11 @@ describe('Rest', function() {
 
     it('Should update with body (as json), criteria and options.', function(done) {
       let injectTest = container.get(InjectTest);
+
+      let responseOutput = {
+        response: null
+      };
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -215,6 +319,10 @@ describe('Rest', function() {
             expect(y.query.sort[0]).toBe(criteriaWithArray.sort[0]);
             expect(y.query.sort[1]).toBe(criteriaWithArray.sort[1]);
             expect(y.originalUrl).toBe(encodeURI(`/posts/?sort[]=${criteriaWithArray.sort[0]}&sort[]=${criteriaWithArray.sort[1]}`));
+          }),
+        injectTest.apiEndpoint.update('posts/', criteriaWithArray, body, options, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
           })
       ]).then(x => {
         done();
@@ -223,6 +331,7 @@ describe('Rest', function() {
 
     it('Should update with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -240,6 +349,11 @@ describe('Rest', function() {
   describe('.updateOne()', function() {
     it('Should update with body (as json), criteria and options.', function(done) {
       let injectTest = container.get(InjectTest);
+
+      let responseOutput = {
+        response: null
+      };
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -261,6 +375,10 @@ describe('Rest', function() {
             expect(y.contentType).toMatch(options.headers['Content-Type']);
             expect(y.Authorization).toBe(options.headers['Authorization']);
           }),
+        injectTest.apiEndpoint.updateOne('posts/', 'id', criteria, body, options, responseOutput)
+          .then(y => {
+          expect(responseOutput.response instanceof Response).toBe(true);
+          }),
         injectTest.apiEndpoint.updateOne('posts/', 'id', criteriaWithArray, body)
           .then(y => {
             expect(y.method).toBe('PUT');
@@ -276,6 +394,7 @@ describe('Rest', function() {
 
     it('Should updateOne with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -294,6 +413,7 @@ describe('Rest', function() {
   describe('.patch()', function() {
     it('Should patch with body (as json).', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -321,6 +441,11 @@ describe('Rest', function() {
 
     it('Should patch with body (as json), criteria and options.', function(done) {
       let injectTest = container.get(InjectTest);
+
+      let responseOutput = {
+        response: null
+      };
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -342,6 +467,10 @@ describe('Rest', function() {
             expect(y.contentType).toMatch(options.headers['Content-Type']);
             expect(y.Authorization).toBe(options.headers['Authorization']);
           }),
+        injectTest.apiEndpoint.patch('post/', criteria, body, options, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
+          }),
         injectTest.apiEndpoint.patch('post/', criteriaWithArray, body)
           .then(y => {
             expect(y.method).toBe('PATCH');
@@ -357,6 +486,7 @@ describe('Rest', function() {
 
     it('Should patch with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -375,6 +505,10 @@ describe('Rest', function() {
   describe('.patchOne()', function() {
     it('Should patch with body (as json), id, criteria and options.', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -396,6 +530,10 @@ describe('Rest', function() {
             expect(y.contentType).toMatch(options.headers['Content-Type']);
             expect(y.Authorization).toBe(options.headers['Authorization']);
           }),
+        injectTest.apiEndpoint.patchOne('post/', 'id', criteria, body, options, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
+          }),
         injectTest.apiEndpoint.patchOne('post/', 'id', criteriaWithArray, body)
           .then(y => {
             expect(y.method).toBe('PATCH');
@@ -411,6 +549,7 @@ describe('Rest', function() {
 
     it('Should patchOne with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -429,6 +568,10 @@ describe('Rest', function() {
   describe('.destroy()', function() {
     it('Should destroy with criteria and options.', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -448,6 +591,10 @@ describe('Rest', function() {
             expect(y.query.comment).toBe(criteria.comment);
             expect(y.Authorization).toBe(options.headers['Authorization']);
           }),
+        injectTest.apiEndpoint.destroy('posts/', criteria, options, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
+          }),
         injectTest.apiEndpoint.destroy('posts/', criteriaWithArray)
           .then(y => {
             expect(y.method).toBe('DELETE');
@@ -463,6 +610,7 @@ describe('Rest', function() {
 
     it('Should destroy with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -481,6 +629,10 @@ describe('Rest', function() {
   describe('.destroyOne()', function() {
     it('Should destroy with id, criteria and options.', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = false;
 
       Promise.all([
@@ -500,6 +652,10 @@ describe('Rest', function() {
             expect(y.query.comment).toBe(criteria.comment);
             expect(y.Authorization).toBe(options.headers['Authorization']);
           }),
+        injectTest.apiEndpoint.destroyOne('posts/', 'id', criteria, options, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
+          }),
         injectTest.apiEndpoint.destroyOne('posts/', 'id', criteriaWithArray)
           .then(y => {
             expect(y.method).toBe('DELETE');
@@ -515,6 +671,7 @@ describe('Rest', function() {
 
     it('Should destroyOne with RFC6570 queries.', function(done) {
       let injectTest = container.get(InjectTest);
+
       injectTest.apiEndpoint.useTraditionalUriTemplates = true;
 
       Promise.all([
@@ -533,6 +690,9 @@ describe('Rest', function() {
   describe('.create()', function() {
     it('Should create body (as json).', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
 
       Promise.all([
         injectTest.apiEndpoint.create('posts', body)
@@ -552,6 +712,9 @@ describe('Rest', function() {
 
     it('Should create body (as json) and options.', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
 
       Promise.all([
         injectTest.apiEndpoint.create('posts', body, options)
@@ -567,6 +730,10 @@ describe('Rest', function() {
             expect(y.path).toBe('/posts/');
             expect(y.contentType).toMatch(options.headers['Content-Type']);
             expect(y.Authorization).toBe(options.headers['Authorization']);
+          }),
+        injectTest.apiEndpoint.create('posts/', body, options, responseOutput)
+          .then(y => {
+            expect(responseOutput.response instanceof Response).toBe(true);
           })
       ]).then(x => {
         done();
@@ -575,13 +742,16 @@ describe('Rest', function() {
   });
 
   describe('.post()', function() {
-    it('Should post body (as urlencoded).', function(done) {
+    it('Should post body (as urlencoded) with custom header (x-www-form-urlencoded).', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
 
       Promise.all([
         injectTest.apiEndpoint.post('posts', body, options)
           .then(y => {
-            expect(JSON.stringify(y.body)).toBe(JSON.stringify(y.body));
+            expect(JSON.stringify(y.body)).toBe(JSON.stringify(body));
             expect(y.method).toBe('POST');
             expect(y.path).toBe('/posts');
             expect(y.contentType).toMatch(options.headers['Content-Type']);
@@ -589,8 +759,106 @@ describe('Rest', function() {
           }),
         injectTest.apiEndpoint.post('posts/', body, options)
           .then(y => {
-            expect(JSON.stringify(y.body)).toBe(JSON.stringify(y.body));
+            expect(JSON.stringify(y.body)).toBe(JSON.stringify(body));
             expect(y.path).toBe('/posts/');
+          })
+      ]).then(x => {
+        done();
+      });
+    });
+
+    it('Should post object body (as urlencoded) with registered default header (x-www-form-urlencoded).', function(done) {
+      let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
+      Promise.all([
+        injectTest.urlencodedEndpoint.post('posts', body)
+          .then(y => {
+            expect(y.method).toBe('POST');
+            expect(y.path).toBe('/posts');
+            expect(y.contentType).toMatch(options.headers['Content-Type']);
+            expect(y.Authorization).toBe(options.headers['Authorization']);
+            expect(y.body.message).toBe('some');
+          }),
+        injectTest.urlencodedEndpoint.post('posts/', body)
+          .then(y => {
+            expect(y.path).toBe('/posts/');
+            expect(y.body.message).toBe('some');
+          })
+      ]).then(x => {
+        done();
+      });
+    });
+
+    it('Should post string body as string with registered default header (x-www-form-urlencoded).', function(done) {
+      let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
+      Promise.all([
+        injectTest.urlencodedEndpoint.post('posts', buildQueryString(body))
+          .then(y => {
+            expect(y.method).toBe('POST');
+            expect(y.path).toBe('/posts');
+            expect(y.contentType).toMatch(options.headers['Content-Type']);
+            expect(y.Authorization).toBe(options.headers['Authorization']);
+            expect(y.body.message).toBe('some');
+          }),
+        injectTest.urlencodedEndpoint.post('posts/', buildQueryString(body))
+          .then(y => {
+            expect(y.path).toBe('/posts/');
+            expect(y.body.message).toBe('some');
+          })
+      ]).then(x => {
+        done();
+      });
+    });
+
+    it('Should post object body (as json) with fetchConfig configuration.', function(done) {
+      let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
+      Promise.all([
+        injectTest.fetchConfigEndpoint.post('posts', body)
+          .then(y => {
+            expect(y.method).toBe('POST');
+            expect(y.path).toBe('/posts');
+            expect(y.contentType).toBe(jsonOptions.headers['Content-Type']);
+            expect(y.body.message).toBe('some');
+          }),
+        injectTest.fetchConfigEndpoint.post('posts/', body)
+          .then(y => {
+            expect(y.path).toBe('/posts/');
+            expect(y.body.message).toBe('some');
+          })
+      ]).then(x => {
+        done();
+      });
+    });
+
+    it('Should post string body as string with fetchConfig configuration.', function(done) {
+      let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
+
+      Promise.all([
+        injectTest.fetchConfigEndpoint.post('posts', JSON.stringify(body))
+          .then(y => {
+            expect(y.method).toBe('POST');
+            expect(y.path).toBe('/posts');
+            expect(y.contentType).toBe(jsonOptions.headers['Content-Type']);
+            expect(y.body.message).toBe('some');
+          }),
+        injectTest.fetchConfigEndpoint.post('posts/', JSON.stringify(body))
+          .then(y => {
+            expect(y.path).toBe('/posts/');
+            expect(y.body.message).toBe('some');
           })
       ]).then(x => {
         done();
@@ -599,8 +867,12 @@ describe('Rest', function() {
 
     it('Should post body (as FormData) and options.', function(done) {
       let injectTest = container.get(InjectTest);
+      let responseOutput = {
+        response: null
+      };
 
       let data = new FormData();
+
       data.append('message', 'some');
 
       Promise.all([